final project

+++ ml-agents/mlagents/trainers/bc/policy.py
@@ -81,8 +81,7 @@ class BCPolicy(TFPolicy):
         else:
             feed_dict[self.model.true_action] = mini_batch["actions"]
             feed_dict[self.model.action_masks] = np.ones(
-                (num_sequences, sum(self.brain.vector_action_space_size)),
-                dtype=np.float32,
+                (num_sequences, sum(self.brain.vector_action_space_size))
             )
         if self.use_vec_obs:
             feed_dict[self.model.vector_in] = mini_batch["vector_obs"]
@@ -90,8 +89,6 @@ class BCPolicy(TFPolicy):
             visual_obs = mini_batch["visual_obs%d" % i]
             feed_dict[self.model.visual_in[i]] = visual_obs
         if self.use_recurrent:
-            feed_dict[self.model.memory_in] = np.zeros(
-                [num_sequences, self.m_size], dtype=np.float32
-            )
+            feed_dict[self.model.memory_in] = np.zeros([num_sequences, self.m_size])
         run_out = self._execute_model(feed_dict, self.update_dict)
         return run_out
